move dashboard parameter options logic to own utils file
